Changes TestSchemaProvider as a class inheriting BigQuerySchemaProvider. Adding Implementation for BigQuerySchemaProvider.
